rag_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai>=1.0.0,<2.0.0
langchain>=0.0.300,<0.2.0
langchain-community>=0.0.20,<0.1.0
langchain-openai>=0.0.1,<0.1.0
faiss-cpu==1.9.0.post1
tiktoken>=0.5.0

